Update CountSubstrings.cpp
no changes made

+++ string/CountSubstrings.cpp
@@ -1,3 +1,5 @@
+// solution written by Aman yadav, codechef username aman212yadav
+// problem link https://www.codechef.com/problems/CSUB
 #include<bits/stdc++.h>
 using namespace std;
 int main()
@@ -8,7 +10,7 @@ int main()
     {cin>>n;
     string s;
     cin>>s;
-    int count1=count(s.begin(),s.end(),'1');
+    long long  count1=count(s.begin(),s.end(),'1');
     cout<<count1+(count1*(count1-1)/2)<<"\n";
     }
 }
